BUILD: Compat with Catlab v0.16 and CombinatorialSpaces v0.6.
Also, add compat entries for Julia standard libraries.
